Minkowski fix for SNumber and SSlider
A few issues here with converting from AWT to Swing:

SSLider:

1) Swing calls addPropertyChangeListener prior to the end of super(),
before boundSupport is created

2) propertyCHange for DValue must set dValue immediately, prior to other
event notification.


SNumber:

1) Swing calls addPropertyChangeListener prior to the end of super(),
before boundSupport is created

2) Minkowski needs to updateValueFromText directly

3) I added an option to allow setting of the format as 
SNumber(double d, String format)
because the +=6.3g was putting in blanks on the right both in Java and
JavaScript. 


Minkowski:

1) changed SNumber format from +-6.3g to +-0.3g -- ignore if desired

2) Swing scroll bars fire their model's document change whenever a value
is changed, so if this is left in, every press of a key in the
speedField runs a new graph

// speedField.addPropertyChangeListener(speedSlider);

Apparently with AWT scrollbars, you can change their position, and they
get out of sync with their object. So that must be something that was
fixed in Swing. But it means you get too many event messages in this
case if you add that property listener. We don't need it.

3) speedField_actionPerformed must update speedField.value from text,
then set speeds, both in the slider and the field (the latter, twice)