Circuit progress -- components showing; Java applet bug fixed
Component painting had to be taken care of with a hack in BorderPanel to
force that panel and its enclosing panel to be not opaque. I think I
could fix this other ways, but that worked.

References to a2s.Applet had to be removed, or at least (now that I
think of it) removed or else I would have had to add one more reference
to that class.

CircuitBuilder needed a Java fix for being an applet. It must never have
been embedded on a web page, as far as I can see.

Dialogs are working in this package; have not tested ValueKnob.